:hammer: criado placar com cronometro e pontuação
