--- jogo-oito-react/src/hook/useJogo.ts
@@ -1,135 +1,0 @@
-import { useState, useEffect } from 'react';
-import useCronometro from './useCronometro';
-
-
-
-
-const useJogo = () => {
-  const {
-    guardarRegistro,
-    pararCronometro,
-    reiniciarCronometro, 
-  } = useCronometro();
-
-  const [tabuleiro, setTabuleiro] = useState<number[][]>([
-    [1, 2, 3],
-    [4, 5, 6],
-    [7, 8, 0]
-  ]);
-
-  const [venceu, setVenceu] = useState(false);
-  const [numVitoria, setNumVitoria] = useState(0)
-
-  useEffect(() => {
-    setVenceu(verificarVitoria());
-  
-   }, [tabuleiro]);
-
- 
- 
-   useEffect(() => {
-     embaralharTabuleiro();
-   }, []);
-
-
-   const mover = (linha: number, coluna: number) => {
-    if (!venceu) {
-      const linhaVazia = tabuleiro.findIndex(row => row.includes(0));
-      const colunaVazia = tabuleiro[linhaVazia].indexOf(0);
-      const novaLinha = linhaVazia + linha;
-      const novaColuna = colunaVazia + coluna;
-
-      if (
-        novaLinha >= 0 &&
-        novaLinha < tabuleiro.length &&
-        novaColuna >= 0 &&
-        novaColuna < tabuleiro[novaLinha].length
-      ) {
-        const novoTabuleiro = [...tabuleiro];
-        novoTabuleiro[linhaVazia][colunaVazia] = tabuleiro[novaLinha][novaColuna];
-        novoTabuleiro[novaLinha][novaColuna] = 0;
-        setTabuleiro(novoTabuleiro);
-      }
-    }
-  };
-
-  const verificarVitoria = () => {   
-    let count = 1;
-    for (let i = 0; i < tabuleiro.length; i++) {
-      for (let j = 0; j < tabuleiro[i].length; j++) {
-        if (tabuleiro[i][j] !== count % 9) {
-          return false;
-        }
-        count++;
-      }
-    }   
-
-    const vitoria = numVitoria + 1
-      setNumVitoria(vitoria)
-      pararCronometro()
-      guardarRegistro()
-   
-    return true;
-  }; 
-
-/*   const verificarVitoria = () => {
-    let count = 1;
-    for (let i = 0; i < tabuleiro.length; i++) {
-      for (let j = 0; j < tabuleiro[i].length; j++) {
-        if (i === tabuleiro.length - 1 && j === tabuleiro[i].length - 1) {
-          if (tabuleiro[i][j] !== 0) {
-            return false;
-          }
-        } else if (tabuleiro[i][j] !== count) {
-          return false;
-        }
-        count++;
-      }
-    }
-    const vitoria = numVitoria + 1;
-    setNumVitoria(vitoria);
-
-    pararCronometro();
-    guardarRegistro();
-  
-    return true;
-  };
-   */
-  const embaralharTabuleiro = () => {
-    const novoTabuleiro = [
-      [1, 2, 3],
-      [4, 5, 6],
-      [7, 8, 0]
-    ];
-    for (let i = novoTabuleiro.length - 1; i > 0; i--) {
-      for (let j = novoTabuleiro[i].length - 1; j > 0; j--) {
-        const k = Math.floor(Math.random() * (i + 1));
-        const l = Math.floor(Math.random() * (j + 1));
-        const temp = novoTabuleiro[i][j];
-        novoTabuleiro[i][j] = novoTabuleiro[k][l];
-        novoTabuleiro[k][l] = temp;
-      }
-    }
-    setTabuleiro(novoTabuleiro);
-  };
- 
-  const reiniciarJogo = () => {  
-
-    embaralharTabuleiro()
-    reiniciarCronometro()
-    
-    setVenceu(false);
-  };
-
-  return {
-    mover,
-    venceu,
-    numVitoria,
-    tabuleiro,
-    verificarVitoria,
-    embaralharTabuleiro,
-    reiniciarJogo,
-  };
-};
-
-export default useJogo;
